+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Added 
+
+- Add `Walking` entity - [#88](https://github.com/IronFoundation/iOS-Woof/pull/88)
+
+### Deleted
+
+- Delete `expired` status from enum `WalkingStatus` [#91](https://github.com/IronFoundation/iOS-Woof/pull/91)
+
 ## [0.5.0] - 27.12.2023
 
 ### Added